fix: return rightly formatted color from AdjustHsl
