+++ RpFilter/Settings.lua
@@ -4,7 +4,7 @@
 import "Turbine.UI.Lotro"
 
 import "Dandiron.RpFilter.ColorPicker"
-import "Dandiron.RpFilter.Color"
+import "Dandiron.RpFilter.ColorUtils"
 
 Settings = {}
 
@@ -207,21 +207,9 @@ local function createSaveButton(colorPickerWindow)
     return saveButton
 end
 
-local function clamp(val, min, max)
-    return math.max(min, math.min(val, max))
-end
-
-local function adjustHsl(color, delta)
-    local hsl = RgbToHsl({r = color.red, g = color.green, b = color.blue})
-    hsl.h = (hsl.h + (delta.h or 0)) % 1
-    hsl.s = clamp(hsl.s + (delta.s or 0), 0, 1)
-    hsl.l = clamp(hsl.l + (delta.l or 0), 0, 1)
-    return HslToRgb(hsl)
-end
-
 local function contrastColors()
-    Settings:getMutable().lighter = adjustHsl(Settings:get().emoteColor, {h = -0.014, l = 0.01})
-    Settings:getMutable().darker = adjustHsl(Settings:get().emoteColor, {h = 0.014, l = -0.008})
+    Settings:getMutable().lighter = AdjustHsl(Settings:get().emoteColor, {h = -0.014, l = 0.01})
+    Settings:getMutable().darker = AdjustHsl(Settings:get().emoteColor, {h = 0.014, l = -0.008})
 end
 
 local function getNewColor(colorPicker)
